requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
pyttsx3==2.90
numpy==1.24.3
pillow==10.0.1
easyocr==1.7.0